+++ src/arkreactor/Compiler/NameResolution/NameResolutionPass.cpp
@@ -19,6 +19,7 @@ namespace Ark::internal
         m_language_symbols.emplace(Language::And);
         m_language_symbols.emplace(Language::Or);
         m_language_symbols.emplace(Language::SysArgs);
+        m_language_symbols.emplace(Language::SysProgramName);
     }
 
     void NameResolutionPass::process(const Node& ast)