+++ src/arkreactor/VM/State.cpp
@@ -25,6 +25,8 @@ namespace Ark
         // default value for builtin__sys:args is empty list
         const Value val(ValueType::List);
         m_binded[std::string(internal::Language::SysArgs)] = val;
+
+        m_binded[std::string(internal::Language::SysProgramName)] = Value("");
     }
 
     bool State::feed(const std::string& bytecode_filename)
@@ -74,30 +76,30 @@ namespace Ark
         return true;
     }
 
-    bool State::doFile(const std::string& file, const uint16_t features)
+    bool State::doFile(const std::string& file_path, const uint16_t features)
     {
-        if (!Utils::fileExists(file))
+        if (!Utils::fileExists(file_path))
         {
-            fmt::print(fmt::fg(fmt::color::red), "Can not find file '{}'\n", file);
+            fmt::print(fmt::fg(fmt::color::red), "Can not find file '{}'\n", file_path);
             return false;
         }
-        m_filename = file;
+        m_filename = file_path;
+        m_binded[std::string(internal::Language::SysProgramName)] = Value(std::filesystem::path(m_filename).filename().string());
 
-        const bytecode_t bytecode = Utils::readFileAsBytes(file);
+        const bytecode_t bytecode = Utils::readFileAsBytes(file_path);
         BytecodeReader bcr;
         bcr.feed(bytecode);
         if (!bcr.checkMagic())  // couldn't read magic number, it's a source file
         {
             // check if it's in the arkscript cache
-            const std::string short_filename = (std::filesystem::path(file)).filename().string();
-            const std::string filename = short_filename.substr(0, short_filename.find_last_of('.')) + ".arkc";
-            const std::filesystem::path directory = (std::filesystem::path(file)).parent_path() / ARK_CACHE_DIRNAME;
-            const std::string path = (directory / filename).string();
+            const std::string filename = std::filesystem::path(file_path).filename().replace_extension(".arkc").string();
+            const std::filesystem::path cache_directory = std::filesystem::path(file_path).parent_path() / ARK_CACHE_DIRNAME;
+            const std::string bytecode_path = (cache_directory / filename).string();
 
-            if (!exists(directory))  // create ark cache directory
-                create_directory(directory);
+            if (!exists(cache_directory))
+                create_directory(cache_directory);
 
-            if (compile(file, path, features) && feed(path))
+            if (compile(file_path, bytecode_path, features) && feed(bytecode_path))
                 return true;
         }
         else if (feed(bytecode))  // it's a bytecode file
@@ -212,6 +214,8 @@ namespace Ark
         // default value for builtin__sys:args is empty list
         const Value val(ValueType::List);
         m_binded[std::string(internal::Language::SysArgs)] = val;
+
+        m_binded[std::string(internal::Language::SysProgramName)] = Value("");
     }
 }